Milk-Tracking-App/api/pyproject.toml
[project]
name = "milk-tracking-api"
version = "0.1.0"
description = "Milk Tracking Application API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.0.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pwdlib[argon2]>=0.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
